// web/src/shared/api/generated/model/runtimeSettingResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { JsonElement } from './jsonElement';
import type { RuntimeSettingGroup } from './runtimeSettingGroup';
import type { RuntimeSettingType } from './runtimeSettingType';

export interface RuntimeSettingResult {
  key: string;
  group: RuntimeSettingGroup;
  type: RuntimeSettingType;
  value: JsonElement;
  defaultValue: JsonElement;
  isOverridden: boolean;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  minimum: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  maximum: number | string | null;
  allowedValues: string[];
  labelArabic: string;
  labelEnglish: string;
  descriptionArabic: string;
  descriptionEnglish: string;
  /** @nullable */
  updatedAt: string | null;
}